requirements.txt
python-telegram-bot==21.9
requests==2.31.0
python-dotenv==1.0.0
validators==0.22.0